Fix ScrollContainer scroll buttons not reflecting overflow immediately
Extract scroll overflow tracking into a custom hook to improve reusability and ensure overflow is detected immediately when content changes. Add test ID to ScrollContainer for better e2e testability.
